+++ apps/web/src/components/write/IntroduceForm.tsx
@@ -1,6 +1,6 @@
 import { Input, InputTemplate, Label, Text, Textarea } from "@moija/ui";
 import { useFormContext } from "react-hook-form";
-import { FormData } from "./page";
+import { FormData } from "../../types/FormData";
 
 export const IntroduceForm = () => {
   const formMethod = useFormContext<FormData>();
