+++ apps/web/src/components/write/QualificationForm.tsx
@@ -1,6 +1,6 @@
 import { ArrowDown, Delete, Input, InputTemplate, Label, Plus, Text } from "@moija/ui";
 import { useFieldArray, useFormContext } from "react-hook-form";
-import { FormData } from "./page";
+import { FormData } from "../../types/FormData";
 
 export const QualificationForm = () => {
   const { control, register } = useFormContext<FormData>();
